vibro: allow lithium charge circuit to control LED pin.
On the board, the LED pin is shared between the lithium charge circuit and the MCU.
This is pretty stupid, given that there are two LEDs on the board, but that's how the HW is.

This change makes it so that the LED pin is changed to an input pin when the MCU isn't driving it (=sleeping), so that the charge circuit gets a chance to drive it.
The penalty of not doing that is, of course, short circuiting the LED output pin of the charge circuit and making it overheat.